fix attention dtype mismatch under bf16 mixed precision
F.softmax can upcast bf16 input to fp32 for numerical stability, leaving
`attn` in fp32 while `v` stays in bf16. The subsequent `torch.matmul(attn, v)`
then raises `RuntimeError: expected scalar type Float but found BFloat16`
when the model is wrapped in FSDP with bf16 MixedPrecision (or run with
`torch.autocast(dtype=bf16)`).

Cast `attn` to `v.dtype` immediately before the matmul in both GQAttention
and MLAttention. In fp32 this is a no-op; in bf16 it re-aligns the dtypes.

+++ open_mythos/main.py
@@ -243,7 +243,7 @@ class GQAttention(nn.Module):
         if mask is not None:
             attn = attn + mask
         attn = self.attn_drop(F.softmax(attn, dim=-1))
-        out = torch.matmul(attn, v)
+        out = torch.matmul(attn.to(v.dtype), v)  # softmax may upcast under bf16 AMP; align with v
         out = out.transpose(1, 2).contiguous().view(B, T, -1)
         return self.wo(out)
 
@@ -385,7 +385,7 @@ class MLAttention(nn.Module):
         if mask is not None:
             attn = attn + mask
         attn = self.attn_drop(F.softmax(attn, dim=-1))
-        out = torch.matmul(attn, v)  # (B, H, T, v_dim)
+        out = torch.matmul(attn.to(v.dtype), v)  # (B, H, T, v_dim); softmax may upcast under bf16 AMP
         out = out.transpose(1, 2).contiguous().view(B, T, -1)
         return self.wo(out)
